Implements iron-a11y-announcer
In order to announce text on screen readers, you need to have a "visible" element in the DOM that changes text content in a specific way. This element is intended to be an a11y announcer on behalf of all other elements that may want to announce text via a screen reader.